% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_preloaded_data}
\alias{load_preloaded_data}
\title{Load Preloaded loaded_data Module}
\usage{
load_preloaded_data(input, output, session, loaded_data_rv, dds_rv)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{loaded_data_rv}{A list  where the loaded_data will be stored. It should contain the following elements:
\itemize{
\item \code{counts}: Raw count loaded_data.
\item \code{samples}: Sample metaloaded_data.
\item \code{norm_counts}: Normalized counts (rlog-transformed).
\item \code{dds}: The DESeq2 loaded_dataset object.
}}

\item{dds_rv}{A reactiveVal container for storing the DESeq2 loaded_dataset.}
}
\value{
A list with elements: counts, samples, norm_counts, dds
}
\description{
This module loads the preprocessed RNASeq loaded_data (counts, samples, normalized counts, and DESeq2 object)
from \code{.rda} files located in the \verb{inst/extloaded_data/} directory. It assigns the loaded loaded_data to the reactive values.
}
