% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_select.R
\name{mod_sample_select}
\alias{mod_sample_select}
\title{Sample Selection Module}
\usage{
mod_sample_select(
  input,
  output,
  session,
  dds_rv,
  loaded_data_rv,
  filtered_data_rv,
  filtered_dds_rv
)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dds_rv}{reactive value full dds object}

\item{loaded_data_rv}{list storing raw_counts samples normailized counts and species}

\item{filtered_data_rv}{a reactive data with filtered tables.}

\item{filtered_dds_rv}{a reactive filtered dds object}
}
\value{
Updates reactive values: \code{filtered_data_rv} (filtered data) and \code{filtered_dds_rv} (filtered DESeq2 object)
}
\description{
This module enables selection of samples from uploaded or demo data.
Users can view the design table and choose samples for downstream analyses.
}
