% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tf.R
\name{mod_tf_enrichment_analysis}
\alias{mod_tf_enrichment_analysis}
\title{Transcription Factor Enrichment Analysis Module}
\usage{
mod_tf_enrichment_analysis(
  input,
  output,
  session,
  geneList_rv,
  processed_tf_data,
  tf_enrichment_result
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{geneList_rv}{Reactive value containing the gene list (log2FC vector) for enrichment}

\item{processed_tf_data}{Reactive value containing selected tf data source data}

\item{tf_enrichment_result}{reactive data containing the tf enrichment results}
}
\description{
This module allows users to perform transcription factor enrichment analysis
based on selected datasets stored locally (gzipped) in inst/extdata.
}
