% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_log.R
\name{mod_logger_server}
\alias{mod_logger_server}
\title{Logger Module Server}
\usage{
mod_logger_server(id, input_all)
}
\arguments{
\item{id}{Shiny module id}

\item{input_all}{The full Shiny input object (passed from main server)}
}
\value{
None. Updates reactive user log internally.
}
\description{
Logger Module Server
}
